120 Wantage Close
"Our quest is practically finished. I shall call with the King to-morrow, and with you, if you care
to come with us. We will be shown into the sitting-room to wait for the lady, but it is probable
that when she comes she may find neither us nor the photograph. It might be a satisfaction to his
Majesty to regain it with his own hands."

"And when will you call?"

"At eight in the morning. She will not be up, so that we shall have a clear field. Besides, we must
be prompt, for this marriage may mean a complete change in her life and habits. I must wire to the
King without delay."

We had reached Baker Street and had stopped at the door. He was searching his pockets for the key
when someone passing said: