1 Glendale
"'I wish to the Lord, Mr. Wilson, that I was a red-headed man.'

"'Why that?' I asks.

"'Why,' says he, 'here's another vacancy on the League of the Red-headed Men. It's worth quite a
little fortune to any man who gets it, and I understand that there are more vacancies than there are
men, so that the trustees are at their wits' end what to do with the money. If my hair would only
change colour, here's a nice little crib all ready for me to step into.'

"'Why, what is it, then?' I asked. You see, Mr. Holmes, I am a very stay-at-home man, and as my
business came to me instead of my having to go to it, I was often weeks on end without putting my
foot over the door-mat. In that way I didn't know much of what was going on outside, and I was
always glad of a bit of news.

"'Have you never heard of the League of the Red-headed Men?' he asked with his eyes open.

"'Never.'