149 Beaconside
"Many times; but what was a fine to me?"

"It must stop here, however," said Bradstreet. "If the police are to hush this thing up, there must
be no more of Hugh Boone."

"I have sworn it by the most solemn oaths which a man can take."

"In that case I think that it is probable that no further steps may be taken. But if you are found
again, then all must come out. I am sure, Mr. Holmes, that we are very much indebted to you for
having cleared the matter up. I wish I knew how you reach your results."

"I reached this one," said my friend, "by sitting upon five pillows and consuming an ounce of shag.
I think, Watson, that if we drive to Baker Street we shall just be in time for breakfast."

I had called upon my friend Sherlock Holmes upon the second morning after Christmas, with the
intention of wishing him the compliments of the season. He was lounging upon the sofa in a purple
dressing-gown, a pipe-rack within his reach upon the right, and a pile of crumpled morning papers,
evidently newly studied, near at hand. Beside the couch was a wooden chair, and on the angle of the
back hung a very seedy and disreputable hard-felt hat, much the worse for wear, and cracked in
several places. A lens and a forceps lying upon the seat of the chair suggested that the hat had
been suspended in this manner for the purpose of examination.