47 Larch Walk
"And now?" I asked.

"Our quest is practically finished. I shall call with the King to-morrow, and with you, if you care
to come with us. We will be shown into the sitting-room to wait for the lady, but it is probable
that when she comes she may find neither us nor the photograph. It might be a satisfaction to his
Majesty to regain it with his own hands."