65 Wantage Close
"He knows it already. When I had cleared it all up I had an interview with him, and finding that he
would not tell me the story, I told it to him, on which he had to confess that I was right and to
add the very few details which were not yet quite clear to me. Your news of this morning, however,
may open his lips."

"For heaven's sake, tell me, then, what is this extraordinary mystery!"

"I will do so, and I will show you the steps by which I reached it. And let me say to you, first,
that which it is hardest for me to say and for you to hear: there has been an understanding between
Sir George Burnwell and your niece Mary. They have now fled together."

"My Mary? Impossible!"

"It is unfortunately more than possible; it is certain. Neither you nor your son knew the true
character of this man when you admitted him into your family circle. He is one of the most dangerous
men in England--a ruined gambler, an absolutely desperate villain, a man without heart or
conscience. Your niece knew nothing of such men. When he breathed his vows to her, as he had done to
a hundred before her, she flattered herself that she alone had touched his heart. The devil knows
best what he said, but at least she became his tool and was in the habit of seeing him nearly every
evening."