25 Tamworth Drive
"And over here is what appears to be the fragment of a hotel bill, which interests me deeply."

"There's nothing in it. I looked at it before," said Lestrade.

"'Oct. 4th, rooms 8s., breakfast 2s. 6d., cocktail 1s., lunch 2s. 6d., glass sherry, 8d.' I see
nothing in that."

"Very likely not. It is most important, all the same. As to the note, it is important also, or at
least the initials are, so I congratulate you again."