119 Wantage Close
"'To Eyford, in Berkshire. It is a little place near the borders of Oxfordshire, and within seven
miles of Reading. There is a train from Paddington which would bring you there at about 11.15.'

"'Very good.'

"'I shall come down in a carriage to meet you.'

"'There is a drive, then?'