93 Wilson Gardens
"I have ordered a carriage," said Lestrade as we sat over a cup of tea. "I knew your energetic
nature, and that you would not be happy until you had been on the scene of the crime."

"It was very nice and complimentary of you," Holmes answered. "It is entirely a question of
barometric pressure."

Lestrade looked startled. "I do not quite follow," he said.

"How is the glass? Twenty-nine, I see. No wind, and not a cloud in the sky. I have a caseful of
cigarettes here which need smoking, and the sofa is very much superior to the usual country hotel
abomination. I do not think that it is probable that I shall use the carriage to-night."