89 Corbet Close
"Surely it would not be difficult to find out. This is not such a populous neighbourhood."

Lestrade shrugged his shoulders. "I am a practical man," he said, "and I really cannot undertake to
go about the country looking for a left-handed gentleman with a game leg. I should become the
laughing-stock of Scotland Yard."