68 Wantage Close
Sherlock Holmes was wrong in his conjecture, however, for there came a step in the passage and a
tapping at the door. He stretched out his long arm to turn the lamp away from himself and towards
the vacant chair upon which a newcomer must sit.

"Come in!" said he.

The man who entered was young, some two-and-twenty at the outside, well-groomed and trimly clad,
with something of refinement and delicacy in his bearing. The streaming umbrella which he held in
his hand, and his long shining waterproof told of the fierce weather through which he had come. He
looked about him anxiously in the glare of the lamp, and I could see that his face was pale and his
eyes heavy, like those of a man who is weighed down with some great anxiety.

"I owe you an apology," he said, raising his golden pince-nez to his eyes. "I trust that I am not
intruding. I fear that I have brought some traces of the storm and rain into your snug chamber."