97 Harvey Street
"I searched the Dundee records, and when I found that the barque Lone Star was there in January,
'85, my suspicion became a certainty. I then inquired as to the vessels which lay at present in the
port of London."

"Yes?"

"The Lone Star had arrived here last week. I went down to the Albert Dock and found that she had
been taken down the river by the early tide this morning, homeward bound to Savannah. I wired to
Gravesend and learned that she had passed some time ago, and as the wind is easterly I have no doubt
that she is now past the Goodwins and not very far from the Isle of Wight."

"What will you do, then?"