56 Larch Walk
"On entering the house, however, I examined, as you remember, the sill and framework of the hall
window with my lens, and I could at once see that someone had passed out. I could distinguish the
outline of an instep where the wet foot had been placed in coming in. I was then beginning to be
able to form an opinion as to what had occurred. A man had waited outside the window; someone had
brought the gems; the deed had been overseen by your son; he had pursued the thief; had struggled
with him; they had each tugged at the coronet, their united strength causing injuries which neither
alone could have effected. He had returned with the prize, but had left a fragment in the grasp of
his opponent. So far I was clear. The question now was, who was the man and who was it brought him
the coronet?

"It is an old maxim of mine that when you have excluded the impossible, whatever remains, however
improbable, must be the truth. Now, I knew that it was not you who had brought it down, so there
only remained your niece and the maids. But if it were the maids, why should your son allow himself
to be accused in their place? There could be no possible reason. As he loved his cousin, however,
there was an excellent explanation why he should retain her secret--the more so as the secret was a
disgraceful one. When I remembered that you had seen her at that window, and how she had fainted on
seeing the coronet again, my conjecture became a certainty.