60 Wellswood Gardens
"Well, we will take it as a working hypothesis for want of a better. Boone, as I have told you, was
arrested and taken to the station, but it could not be shown that there had ever before been
anything against him. He had for years been known as a professional beggar, but his life appeared to
have been a very quiet and innocent one. There the matter stands at present, and the questions which
have to be solved--what Neville St. Clair was doing in the opium den, what happened to him when
there, where is he now, and what Hugh Boone had to do with his disappearance--are all as far from a
solution as ever. I confess that I cannot recall any case within my experience which looked at the
first glance so simple and yet which presented such difficulties."

While Sherlock Holmes had been detailing this singular series of events, we had been whirling
through the outskirts of the great town until the last straggling houses had been left behind, and
we rattled along with a country hedge upon either side of us. Just as he finished, however, we drove
through two scattered villages, where a few lights still glimmered in the windows.

"We are on the outskirts of Lee," said my companion. "We have touched on three English counties in
our short drive, starting in Middlesex, passing over an angle of Surrey, and ending in Kent. See
that light among the trees? That is The Cedars, and beside that lamp sits a woman whose anxious ears
have already, I have little doubt, caught the clink of our horse's feet."

"But why are you not conducting the case from Baker Street?" I asked.

"Because there are many inquiries which must be made out here. Mrs. St. Clair has most kindly put
two rooms at my disposal, and you may rest assured that she will have nothing but a welcome for my
friend and colleague. I hate to meet her, Watson, when I have no news of her husband. Here we are.
Whoa, there, whoa!"