93 Glan Road
"This went on day after day, Mr. Holmes, and on Saturday the manager came in and planked down four
golden sovereigns for my week's work. It was the same next week, and the same the week after. Every
morning I was there at ten, and every afternoon I left at two. By degrees Mr. Duncan Ross took to
coming in only once of a morning, and then, after a time, he did not come in at all. Still, of
course, I never dared to leave the room for an instant, for I was not sure when he might come, and
the billet was such a good one, and suited me so well, that I would not risk the loss of it.

"Eight weeks passed away like this, and I had written about Abbots and Archery and Armour and
Architecture and Attica, and hoped with diligence that I might get on to the B's before very long.
It cost me something in foolscap, and I had pretty nearly filled a shelf with my writings. And then
suddenly the whole business came to an end."

"To an end?"