103 Beaconside
"No, no; it is, it is his very own writing!"

"Very well. It may, however, have been written on Monday and only posted to-day."

"That is possible."

"If so, much may have happened between."

"Oh, you must not discourage me, Mr. Holmes. I know that all is well with him. There is so keen a
sympathy between us that I should know if evil came upon him. On the very day that I saw him last he
cut himself in the bedroom, and yet I in the dining-room rushed upstairs instantly with the utmost
certainty that something had happened. Do you think that I would respond to such a trifle and yet be
ignorant of his death?"