48 Tamworth Drive
"Yes. It was the bisulphate of baryta."

"No, no, the mystery!" I cried.

"Oh, that! I thought of the salt that I have been working upon. There was never any mystery in the
matter, though, as I said yesterday, some of the details are of interest. The only drawback is that
there is no law, I fear, that can touch the scoundrel."

"Who was he, then, and what was his object in deserting Miss Sutherland?"