118 Harvey Street
"I came to find a friend."

"And I to find an enemy."

"An enemy?"

"Yes; one of my natural enemies, or, shall I say, my natural prey. Briefly, Watson, I am in the
midst of a very remarkable inquiry, and I have hoped to find a clue in the incoherent ramblings of
these sots, as I have done before now. Had I been recognised in that den my life would not have been
worth an hour's purchase; for I have used it before now for my own purposes, and the rascally Lascar
who runs it has sworn to have vengeance upon me. There is a trap-door at the back of that building,
near the corner of Paul's Wharf, which could tell some strange tales of what has passed through it
upon the moonless nights."

"What! You do not mean bodies?"

"Ay, bodies, Watson. We should be rich men if we had £1000 for every poor devil who has been done to
death in that den. It is the vilest murder-trap on the whole riverside, and I fear that Neville St.
Clair has entered it never to leave it more. But our trap should be here." He put his two
forefingers between his teeth and whistled shrilly--a signal which was answered by a similar whistle
from the distance, followed shortly by the rattle of wheels and the clink of horses' hoofs.