26 Valentia Road
"Deserted you?"

"Yes. Her bed this morning had not been slept in, her room was empty, and a note for me lay upon the
hall table. I had said to her last night, in sorrow and not in anger, that if she had married my boy
all might have been well with him. Perhaps it was thoughtless of me to say so. It is to that remark
that she refers in this note:

"'My dearest Uncle: "'I feel that I have brought trouble upon you, and that if I had acted
differently this terrible misfortune might never have occurred. I cannot, with this thought in my
mind, ever again be happy under your roof, and I feel that I must leave you forever. Do not worry
about my future, for that is provided for; and, above all, do not search for me, for it will be
fruitless labour and an ill-service to me. In life or in death, I am ever "'Your loving "'Mary.'

"What could she mean by that note, Mr. Holmes? Do you think it points to suicide?"