165 Beaconside
There was little difficulty in entering the grounds, for unrepaired breaches gaped in the old park
wall. Making our way among the trees, we reached the lawn, crossed it, and were about to enter
through the window when out from a clump of laurel bushes there darted what seemed to be a hideous
and distorted child, who threw itself upon the grass with writhing limbs and then ran swiftly across
the lawn into the darkness.

"My God!" I whispered; "did you see it?"

Holmes was for the moment as startled as I. His hand closed like a vice upon my wrist in his
agitation. Then he broke into a low laugh and put his lips to my ear.

"It is a nice household," he murmured. "That is the baboon."

I had forgotten the strange pets which the doctor affected. There was a cheetah, too; perhaps we
might find it upon our shoulders at any moment. I confess that I felt easier in my mind when, after
following Holmes' example and slipping off my shoes, I found myself inside the bedroom. My companion
noiselessly closed the shutters, moved the lamp onto the table, and cast his eyes round the room.
All was as we had seen it in the daytime. Then creeping up to me and making a trumpet of his hand,
he whispered into my ear again so gently that it was all that I could do to distinguish the words: