95 Harvey Street
"In what way?" asked Holmes.

"It is no time for me to hide anything. James and his father had many disagreements about me. Mr.
McCarthy was very anxious that there should be a marriage between us. James and I have always loved
each other as brother and sister; but of course he is young and has seen very little of life yet,
and--and--well, he naturally did not wish to do anything like that yet. So there were quarrels, and
this, I am sure, was one of them."