39 Harvey Street
"Tell me," broke in the engineer, "is Dr. Becher a German, very thin, with a long, sharp nose?"

The station-master laughed heartily. "No, sir, Dr. Becher is an Englishman, and there isn't a man in
the parish who has a better-lined waistcoat. But he has a gentleman staying with him, a patient, as
I understand, who is a foreigner, and he looks as if a little good Berkshire beef would do him no
harm."