169 Larch Walk
"My dear fellow, there lies the problem. It is true that 'For Mrs. Henry Baker' was printed upon a
small card which was tied to the bird's left leg, and it is also true that the initials 'H. B.' are
legible upon the lining of this hat, but as there are some thousands of Bakers, and some hundreds of
Henry Bakers in this city of ours, it is not easy to restore lost property to any one of them."

"What, then, did Peterson do?"

"He brought round both hat and goose to me on Christmas morning, knowing that even the smallest
problems are of interest to me. The goose we retained until this morning, when there were signs
that, in spite of the slight frost, it would be well that it should be eaten without unnecessary
delay. Its finder has carried it off, therefore, to fulfil the ultimate destiny of a goose, while I
continue to retain the hat of the unknown gentleman who lost his Christmas dinner."

"Did he not advertise?"

"No."