21 Wantage Close
"'Not at all. I only doubt--'

"'The propriety of my leaving it. You may set your mind at rest about that. I should not dream of
doing so were it not absolutely certain that I should be able in four days to reclaim it. It is a
pure matter of form. Is the security sufficient?'

"'Ample.'