14 Wilson Gardens
"Oh, it is a matter of no importance; but I don't know why you should be so warm over such a
trifle."

"Warm! You'd be as warm, maybe, if you were as pestered as I am. When I pay good money for a good
article there should be an end of the business; but it's 'Where are the geese?' and 'Who did you
sell the geese to?' and 'What will you take for the geese?' One would think they were the only geese
in the world, to hear the fuss that is made over them."

"Well, I have no connection with any other people who have been making inquiries," said Holmes
carelessly. "If you won't tell us the bet is off, that is all. But I'm always ready to back my
opinion on a matter of fowls, and I have a fiver on it that the bird I ate is country bred."