8 Ventnor Terrace
A loud thudding noise came from somewhere downstairs. "That is Mrs. Toller in the cellar," said she.
"Her husband lies snoring on the kitchen rug. Here are his keys, which are the duplicates of Mr.
Rucastle's."

"You have done well indeed!" cried Holmes with enthusiasm. "Now lead the way, and we shall soon see
the end of this black business."

We passed up the stair, unlocked the door, followed on down a passage, and found ourselves in front
of the barricade which Miss Hunter had described. Holmes cut the cord and removed the transverse
bar. Then he tried the various keys in the lock, but without success. No sound came from within, and
at the silence Holmes' face clouded over.

"I trust that we are not too late," said he. "I think, Miss Hunter, that we had better go in without
you. Now, Watson, put your shoulder to it, and we shall see whether we cannot make our way in."

It was an old rickety door and gave at once before our united strength. Together we rushed into the
room. It was empty. There was no furniture save a little pallet bed, a small table, and a basketful
of linen. The skylight above was open, and the prisoner gone.