91 Balcombe Close
"'Then we can hardly get there before midnight. I suppose there would be no chance of a train back.
I should be compelled to stop the night.'

"'Yes, we could easily give you a shake-down.'

"'That is very awkward. Could I not come at some more convenient hour?'

"'We have judged it best that you should come late. It is to recompense you for any inconvenience
that we are paying to you, a young and unknown man, a fee which would buy an opinion from the very
heads of your profession. Still, of course, if you would like to draw out of the business, there is
plenty of time to do so.'

"I thought of the fifty guineas, and of how very useful they would be to me. 'Not at all,' said I,
'I shall be very happy to accommodate myself to your wishes. I should like, however, to understand a
little more clearly what it is that you wish me to do.'